requirements.txt
#dgl==0.8.0.post1
dgl_cu113==0.8.0.post2
dill==0.3.4
matplotlib==3.5.1
networkx==2.7.1
numpy==1.19.5
openpyxl==3.0.5
pandas==1.1.4
psutil==5.9.0
PyYAML==6.0
ray==1.13.0
scikit_learn==1.1.1
scipy==1.4.1
torch==1.11.0+cu113
tqdm==4.52.0
tensorboard==2.11.0
seaborn==0.11.2